requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
apipkg==1.5
appnope==0.1.2
attrs==20.3.0
backcall==0.2.0
beeprint==2.4.10
decorator==5.0.7
et-xmlfile==1.1.0
execnet==1.8.0
Faker==6.6.3
filelock==3.0.12
importlib-metadata==2.0.0
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
mccabe==0.6.1
openpyxl==3.0.7
packaging==20.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.9.0
pycodestyle==2.7.0
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.1.2
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-sugar==0.9.4
pytest-xdist==2.2.1
python-dateutil==2.8.1
PyYAML==5.3.1
selenium==3.141.0
seletools==1.0.3
six==1.15.0
tblib==1.7.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.0
urwid==2.1.2
wcwidth==0.2.5
zipp==3.4.0
